fix: export module types under pakcage name
