+++ package.json
@@ -75,7 +75,7 @@
     "@shikijs/transformers": "^1.6.4",
     "@types/hast": "^3.0.4",
     "@types/mdast": "^4.0.4",
-    "@vue/compiler-core": "^3.4.27",
+    "@vue/compiler-core": "^3.4.28",
     "consola": "^3.2.3",
     "debug": "^4.3.5",
     "defu": "^6.1.4",
@@ -113,7 +113,8 @@
     "@nuxt/module-builder": "^0.7.1",
     "@nuxt/schema": "^3.12.1",
     "@nuxt/test-utils": "^3.13.1",
-    "@nuxt/ui": "^2.16.0",
+    "@nuxt/ui": "^2.17.0",
+    "@nuxtjs/mdc": "link:.",
     "@types/node": "^20.14.2",
     "changelogen": "^0.5.5",
     "eslint": "^9.4.0",
