+++ src/config.ts
@@ -1,5 +1,1 @@
-import type { MdcConfig } from './types/config'
-
-export function defineConfig(config: MdcConfig) {
-  return config
-}
+export { defineConfig } from './module'
