+++ src/module.ts
@@ -3,12 +3,16 @@ import { defineNuxtModule, extendViteConfig, addComponent, addComponentsDir, cre
 import { defu } from 'defu'
 import { resolve } from 'pathe'
 import type { BundledLanguage } from 'shiki'
-import type { ModuleOptions } from './types/module'
+import type { ModuleOptions, MdcConfig } from './types'
 import { registerMDCSlotTransformer } from './utils/vue-mdc-slot'
 import * as templates from './templates'
 import { addWasmSupport } from './utils'
 
-export * from './types/module'
+export type * from './types'
+
+export function defineConfig(config: MdcConfig) {
+  return config
+}
 
 export const DefaultHighlightLangs: BundledLanguage[] = [
   'js',
