+++ src/runtime/components/MDCRenderer.vue
@@ -4,8 +4,8 @@ import destr from 'destr'
 import { kebabCase, pascalCase } from 'scule'
 import { find, html } from 'property-information'
 import type { VNode, ConcreteComponent, PropType, DefineComponent } from 'vue'
+import type { MDCElement, MDCNode, MDCRoot, MDCData } from '@nuxtjs/mdc'
 import htmlTags from '../parser/utils/html-tags-list'
-import type { MDCElement, MDCNode, MDCRoot, MDCData } from '../../types'
 import { flatUnwrap } from '../utils/node'
 
 type CreateElement = typeof h
