+++ src/runtime/highlighter/rehype.ts
@@ -1,12 +1,7 @@
 import type { Root, Element } from 'hast'
 import { visit } from 'unist-util-visit'
 import { toString } from 'hast-util-to-string'
-import type { Highlighter, MdcThemeOptions } from '../../types/highlighter'
-
-export interface RehypeHighlightOption {
-  theme?: MdcThemeOptions
-  highlighter?: Highlighter
-}
+import type { RehypeHighlightOption } from '@nuxtjs/mdc'
 
 export default rehypeHighlight
