+++ src/runtime/parser/index.ts
@@ -3,9 +3,8 @@ import remarkParse from 'remark-parse'
 import remark2rehype from 'remark-rehype'
 import { parseFrontMatter } from 'remark-mdc'
 import { defu } from 'defu'
-import type { MDCData, MDCElement, MDCParseOptions, MDCParserResult, MDCRoot, Toc } from '../../types'
+import type { MdcConfig, MDCData, MDCElement, MDCParseOptions, MDCParserResult, MDCRoot, Toc } from '@nuxtjs/mdc'
 import { nodeTextContent } from '../utils/node'
-import type { MdcConfig } from '../../types/config'
 import { useProcessorPlugins } from './utils/plugins'
 import { defaults } from './options'
 import { generateToc } from './toc'
