+++ src/types/index.ts
@@ -1,4 +1,7 @@
-export * from './tree'
-export * from './parser'
-export * from './toc'
-export * from './module'
+export type * from './tree'
+export type * from './parser'
+export type * from './toc'
+export type * from './module'
+export type * from './highlighter'
+export type * from './config'
+export type * from './hast'
